Moved code to get memory_limit in bytes to command functions since that is where it is used.
