+++ src/command_functions.php
@@ -487,3 +487,29 @@ function get_path_to_config(array $options): ?string
     }
     return $path_to_config;
 }
+
+function getMemoryLimitInBytes(): int
+{
+    $limit = ini_get('memory_limit');
+    // for unlimited = -1
+    if ($limit < 0) {
+        return -1;
+    }
+
+    if (preg_match('/^(\d+)(\D?)$/', $limit, $matches)) {
+        $limit = (int)$matches[1];
+        switch (strtoupper($matches[2] ?? '')) {
+            case 'G':
+                $limit *= 1024 * 1024 * 1024;
+                break;
+            case 'M':
+                $limit *= 1024 * 1024;
+                break;
+            case 'K':
+                $limit *= 1024;
+                break;
+        }
+    }
+
+    return (int)$limit;
+}
